# leb128: no allocations in decoding