--- a/internal/leb128/leb128_alloc_test.go
+++ b/internal/leb128/leb128_alloc_test.go
@@ -2,10 +2,12 @@ package leb128
 
 import (
 	"bytes"
-	"github.com/tetratelabs/wazero/internal/testing/require"
 	"testing"
+
+	"github.com/tetratelabs/wazero/internal/testing/require"
 )
 
+// TestLeb128NoAlloc ensures no allocation required in the leb128 package.
 func TestLeb128NoAlloc(t *testing.T) {
 	t.Run("LoadUint32", func(t *testing.T) {
 		result := testing.Benchmark(BenchmarkLoadUint32)