This is a hot pass during compilation. Found while doing #936, and raised this as a separate PR.
```
benchstat leb128_old.txt leb128_new.txt
name             old time/op    new time/op    delta
LoadUint32-10      41.8ns ± 1%    12.0ns ± 0%   -71.35%  (p=0.000 n=10+10)
LoadUint64-10      4.96ns ± 0%    4.95ns ± 0%      ~     (p=0.813 n=9+9)
LoadInt32-10       42.1ns ± 2%    12.4ns ± 0%   -70.56%  (p=0.000 n=10+10)
LoadInt64-10       41.7ns ± 1%    12.4ns ± 0%   -70.27%  (p=0.000 n=10+10)
DecodeUint32-10    38.4ns ± 1%    17.3ns ± 1%   -55.03%  (p=0.000 n=10+10)
DecodeInt32-10     39.3ns ± 1%    17.8ns ± 0%   -54.59%  (p=0.000 n=10+9)
DecodeInt64-10     39.3ns ± 0%    17.6ns ± 0%   -55.28%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
LoadUint32-10       28.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
LoadUint64-10       0.00B          0.00B           ~     (all equal)
LoadInt32-10        28.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
LoadInt64-10        28.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
DecodeUint32-10     16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
DecodeInt32-10      16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)
DecodeInt64-10      16.0B ± 0%      0.0B       -100.00%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
LoadUint32-10        2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
LoadUint64-10        0.00           0.00           ~     (all equal)
LoadInt32-10         2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
LoadInt64-10         2.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
DecodeUint32-10      1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
DecodeInt32-10       1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
DecodeInt64-10       1.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)

```


Signed-off-by: Takeshi Yoneda <takeshi@tetrate.io>